+++ CHANGELOG.md
@@ -1,2 +1,5 @@
-# 1.0.2 Beta
-- Made the Vortex Manipulator have an eye of ender as a temp texture
+# 1.1.0 Beta
+- Added custom texture for Vortex Manipulator
+- Added custom background for the screens
+- Added a durability of 6 for the Manipulator
+- Made the tooltip text red